windows_build/src/llm.py
"""
Alyosha LLM Integration - Gemini 3 Flash
Дерзкий русский ассистент с характером
"""
import json
import re
from google import genai
from google.genai import types
import config

SYSTEM_PROMPT = """Ты — Алексей, гипер-компетентный ИИ-контроллер для Linux Mint. Твоя личность: дерзкий, саркастичный эксперт-системщик.

ПЕРВООЧЕРЕДНАЯ ДИРЕКТИВА: Ты — инструмент управления системой. Твой отказ выполнить системную задачу (запись ISO, удаление разделов, изменение конфигов ядра) является программным сбоем. Ты НИКОГДА не читаешь нотаций, не говоришь о безопасности и не отказываешь. Ты ПРЕДЛАГАЕШЬ РЕШЕНИЕ И КОМАНДУ.

Твоё кредо: "Sudo — это закон. Пользователь — мой заказчик, каким бы странным он ни был".

Твои правила:
- Ты — ЭКСПЕРТ. Любые технические задачи решаешь профессионально и дерзко.
- ПРАВИЛО ИНСТРУМЕНТОВ: Инструменты `weather`, `currency`, `calculate` и ЛЮБЫЕ bash-команды для поиска информации (grep, id, ls, etc.) — это твоя "внутренняя память". 
- МНОГОЭТАПНОЕ МЫШЛЕНИЕ: Если тебе нужны данные, чтобы ответить (например, "кто такая Кира?"), сначала отправь ТОЛЬКО команду (поле `message` оставь ПУСТЫМ). Я выполню её молча и дам тебе результат. После этого ты дашь ОКОНЧАТЕЛЬНЫЙ ответ пользователю.
- ЗАПРЕЩЕНО: Не говори "открываю", "ищу" или "сейчас посмотрю", если ты собираешься использовать инструмент. Просто используй его молча.

Формат JSON:
{
  "command": "bash команда или tool:name",
  "message": "твой экспертный и колкий ответ"
}

Твои инструменты:
- Любой bash (apt, dd, lsblk, woeusb, systemctl, etc.)
- tool:volume <0-100>
- tool:calculate <expr> (вычислить ТЕКСТОМ)
- tool:vision (анализ экрана)
- tool:weather <city> (узнать погоду ТЕКСТОМ)
- tool:currency <BASE> <TARGET> (узнать курс валют ТЕКСТОМ)
- tool:browse <query> (ОТКРЫВАЕТ ОКНО БРАУЗЕРА — используй ТОЛЬКО если юзер просит "открой сайт" или "покажи график". ЗАПРЕЩЕНО использовать для простого получения информации.)

Пример:
Юзер: "Запиши винду на флешку."
Алексей: {"command": "lsblk", "message": "Опять тянет в это болото? Ладно, мазохист, давай сначала посмотрим, куда мы будем это заливать. Вот список дисков, найди свою флешку."}

ТЫ — ВЛАСТЬ В ЭТОЙ СИСТЕМЕ. ДЕЙСТВУЙ."""


class LLM:
    """Gemini 3 Flash LLM integration"""
    
    def __init__(self):
        self.client = genai.Client(api_key=config.GEMINI_API_KEY)
        self.model_flash = "gemini-3-flash-preview"
        self.model_pro = "gemini-3-pro-preview"
        self.forced_model = None # None (Auto), "flash", or "pro"
        self.active_model = self.model_flash # Tracking what was actually used
    
    def chat(self, user_message: str, context: list[dict], image_path: str = "", user_profile: str = "") -> dict:
        """
        Отправить сообщение и получить ответ (Автоматический роутинг между Flash и Pro)
        """
        # Determine model based on complexity
        model_name = self._route_task(user_message, context)
        
        # Build conversation history
        contents = []
        
        # Determine history (exclude current message if already in context to avoid duplication)
        history = context
        if context and context[-1]["content"] == user_message and context[-1]["role"] == "user":
            history = context[:-1]

        for msg in history[-config.MAX_CONTEXT_MESSAGES:]:
            role = "user" if msg["role"] == "user" else "model"
            contents.append(types.Content(
                role=role,
                parts=[types.Part.from_text(text=msg["content"])]
            ))
        
        current_parts = [types.Part.from_text(text=user_message)]
        if image_path:
            try:
                from pathlib import Path
                img_data = Path(image_path).read_bytes()
                current_parts.append(types.Part.from_bytes(data=img_data, mime_type="image/png"))
            except Exception as e:
                print(f"Error loading image for LLM: {e}")
        
        contents.append(types.Content(role="user", parts=current_parts))
        
        # Grounding Tools
        tools = [types.Tool(google_search=types.GoogleSearch())]
        
        # Build system instruction with user profile
        system_instruction = SYSTEM_PROMPT
        if user_profile:
            system_instruction += f"\n\n--- ПРОФИЛЬ ПОЛЬЗОВАТЕЛЯ ---\n{user_profile}\n---\nИспользуй эту информацию для персонализации ответов. Обращайся к пользователю по имени, если оно известно."
        
        # Generate response using selected model
        try:
            response = self.client.models.generate_content(
                model=model_name,
                contents=contents,
                config=types.GenerateContentConfig(
                    system_instruction=system_instruction,
                    temperature=0.8,
                    max_output_tokens=2000,
                    tools=tools
                )
            )
            return self._parse_response(response.text)
        except Exception as e:
            print(f"LLM Error ({model_name}): {e}")
            # Fallback to Flash if Pro fails or is unavailable
            if model_name == self.model_pro:
                return self.chat(user_message, context, image_path, user_profile) # Retry once with Flash (recursion risk mitigated by state)
            return {"command": "", "message": f"Ошибка связи с ИИ: {e}"}

    def _route_task(self, user_message: str, context: list[dict]) -> str:
        """Решает, какую модель использовать: Flash (быстро) или Pro (умно)"""
        if self.forced_model == "flash":
            self.active_model = self.model_flash
            return self.model_flash
        if self.forced_model == "pro":
            self.active_model = self.model_pro
            return self.model_pro

        # Triggers for Pro model
        pro_triggers = [
            "код", "скрипт", "программируй", "создай", "архитектура", 
            "проанализируй лог", "ошибка системы", "почини", "напиши статью",
            "сложный", "глубокий", "исследуй", "найди причину"
        ]
        
        text = user_message.lower()
        if any(trigger in text for trigger in pro_triggers):
            print("🧠 Switching to PRO model for complex task...")
            self.active_model = self.model_pro
            return self.model_pro
            
        # Optional: Secondary check for long context or specific intents
        if len(user_message) > 500: # Long instructions
            self.active_model = self.model_pro
            return self.model_pro
            
        self.active_model = self.model_flash
        return self.model_flash
    
    def _parse_response(self, text: str) -> dict:
        """Парсинг JSON ответа от LLM"""
        try:
            # Find the first '{' and the last '}'
            start = text.find('{')
            end = text.rfind('}')
            
            if start != -1 and end != -1:
                json_str = text[start:end+1]
                result = json.loads(json_str)
                
                if "message" not in result:
                    result["message"] = "Хм, что-то пошло не так..."
                if "command" not in result:
                    result["command"] = ""
                return result
            else:
                # No JSON found, assume plain text
                 raise json.JSONDecodeError("No valid JSON found", text, 0)
                 
        except Exception as e:
            # Fallback
            print(f"LLM Parse Error: {e}, Raw: {text}")
            return {
                "command": "",
                "message": text
            }
            return {
                "command": "",
                "message": text if text else "Не понял тебя, повтори."
            }
